feat(forms): 구매문의 앱 이동
