--- packages/forms/src/index.tsx
@@ -1,13 +1,0 @@
-import React from 'react'
-import { createRoot } from 'react-dom/client'
-import App from './App'
-import GlobalStyle from 'styles/GlobalStyle'
-
-const container = document.getElementById('root')
-const root = createRoot(container!)
-root.render(
-  <React.StrictMode>
-    <GlobalStyle />
-    <App />
-  </React.StrictMode>,
-)
